# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Dekopen API
  version: 1.0.0
  description: Authenticated tenant and engine API boundary.
paths:
...
  /api/v1/positions/{position_id}/:
    get:
      operationId: positions_retrieve
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: position_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    put:
      operationId: positions_update
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: position_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionUpdateRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    delete:
      operationId: positions_destroy
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: query
        name: expected_updated_at
        schema:
          type: string
          format: date-time
        required: true
      - in: path
        name: position_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      security:
      - SupabaseBearer: []
      responses:
        '204':
          description: No response body
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
  /api/v1/projects/:
    get:
      operationId: projects_list
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      tags:
      - projects
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    post:
      operationId: projects_create
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectWriteRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/projects/{project_id}/:
    get:
      operationId: projects_retrieve
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    patch:
      operationId: projects_update
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProjectUpdateRequest'
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/projects/{project_id}/clone/:
    post:
      operationId: projects_clone
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloneProjectRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/projects/{project_id}/positions/:
    post:
      operationId: positions_create
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionWriteRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/projects/{project_id}/successor/:
    post:
      operationId: projects_start_successor
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuccessorRequestRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
components:
  schemas:
...
    CloneProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        expected_updated_at:
          type: string
          format: date-time
      required:
      - expected_updated_at
...
    EngineCalculateRequest:
      type: object
      properties:
        system_id:
          type: string
          format: uuid
        nominal_width_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        nominal_height_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        color:
          type: string
          maxLength: 50
        parametric_tree: {}
      required:
      - color
      - nominal_height_mm
      - nominal_width_mm
      - parametric_tree
      - system_id
...
    EngineCalculateResponse:
      type: object
      properties:
        profile_cuts:
          type: array
          items:
            $ref: '#/components/schemas/ProfileCut'
        reinforcements:
          type: array
          items:
            $ref: '#/components/schemas/Reinforcement'
        glasses:
          type: array
          items:
            $ref: '#/components/schemas/GlassPiece'
        panels:
          type: array
          items:
            $ref: '#/components/schemas/PanelPiece'
        hardware_items:
          type: array
          items:
            $ref: '#/components/schemas/HardwareItem'
        leaf_weights:
          type: array
          items:
            $ref: '#/components/schemas/LeafWeight'
        calculation_hash:
          type: string
          pattern: ^sha256:[0-9a-f]{64}$
      required:
      - calculation_hash
      - glasses
      - hardware_items
      - leaf_weights
      - panels
      - profile_cuts
      - reinforcements
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        required_aal:
          $ref: '#/components/schemas/RequiredAalEnum'
      required:
      - code
      - detail
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/Membership'
      required:
      - error
...
    Membership:
      type: object
      properties:
        organization_id:
          type: string
          format: uuid
        organization_name:
          type: string
        role:
          $ref: '#/components/schemas/RoleEnum'
      required:
      - organization_id
      - organization_name
      - role
...
    PatchedProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        client_name:
          type: string
          minLength: 1
          maxLength: 255
        client_rut:
          type: string
          maxLength: 50
        client_email:
          oneOf:
          - type: string
            format: email
          - type: string
            maxLength: 0
        client_phone:
          type: string
          maxLength: 50
        delivery_address:
          type: string
        notes_commercial:
          type: string
        notes_internal:
          type: string
        expected_updated_at:
          type: string
          format: date-time
...
    PositionDesignRequest:
      type: object
      properties:
        system_id:
          type: string
          format: uuid
        nominal_width_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        nominal_height_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        color:
          $ref: '#/components/schemas/ColorEnum'
        parametric_tree: {}
      required:
      - color
      - nominal_height_mm
      - nominal_width_mm
      - parametric_tree
      - system_id
...
    PositionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        position_index:
          type: integer
        location_tag:
          type: string
          nullable: true
        quantity:
          type: integer
        typology:
          type: string
        design:
          $ref: '#/components/schemas/EngineCalculateRequest'
        bom:
          $ref: '#/components/schemas/EngineCalculateResponse'
        updated_at:
          type: string
          format: date-time
      required:
      - bom
      - design
      - id
      - location_tag
      - position_index
      - project_id
      - quantity
      - typology
      - updated_at
    PositionUpdateRequest:
      type: object
      properties:
        location_tag:
          type: string
          maxLength: 100
        quantity:
          type: integer
          maximum: 2147483647
          minimum: 1
        design:
          $ref: '#/components/schemas/PositionDesignRequest'
        expected_updated_at:
          type: string
          format: date-time
      required:
      - design
      - expected_updated_at
      - location_tag
      - quantity
    PositionWriteRequest:
      type: object
      properties:
        location_tag:
          type: string
          maxLength: 100
        quantity:
          type: integer
          maximum: 2147483647
          minimum: 1
        design:
          $ref: '#/components/schemas/PositionDesignRequest'
      required:
      - design
      - location_tag
      - quantity
...
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
      required:
      - items
    ProjectResponse:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        client_name:
          type: string
          maxLength: 255
        client_rut:
          type: string
          maxLength: 50
        client_email:
          oneOf:
          - type: string
            format: email
          - type: string
            maxLength: 0
        client_phone:
          type: string
          maxLength: 50
        delivery_address:
          type: string
        notes_commercial:
          type: string
        notes_internal:
          type: string
        id:
          type: string
          format: uuid
        code:
          type: string
        status:
          $ref: '#/components/schemas/ProjectResponseStatusEnum'
        current_revision:
          type: string
          pattern: ^REV-[A-Z]+$
        total_price_net:
          type: string
        total_price_tax:
          type: string
        total_price_gross:
          type: string
        pricing_current:
          type: boolean
        current_pricing_operation_id:
          type: string
          format: uuid
          nullable: true
        position_count:
          type: integer
        updated_at:
          type: string
          format: date-time
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionResponse'
        versions:
          type: array
          items:
            $ref: '#/components/schemas/ProjectVersionResponse'
      required:
      - client_name
      - code
      - current_pricing_operation_id
      - current_revision
      - id
      - name
      - position_count
      - pricing_current
      - status
      - total_price_gross
      - total_price_net
      - total_price_tax
      - updated_at
    ProjectResponseStatusEnum:
      enum:
      - DRAFT
      - QUOTED
      - APPROVED
      - IN_PRODUCTION
      - COMPLETED
      - CANCELLED
      type: string
      description: |-
        * `DRAFT` - DRAFT
        * `QUOTED` - QUOTED
        * `APPROVED` - APPROVED
        * `IN_PRODUCTION` - IN_PRODUCTION
        * `COMPLETED` - COMPLETED
        * `CANCELLED` - CANCELLED
...
    ProjectWriteRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        client_name:
          type: string
          minLength: 1
          maxLength: 255
        client_rut:
          type: string
          maxLength: 50
        client_email:
          oneOf:
          - type: string
            format: email
          - type: string
            maxLength: 0
        client_phone:
          type: string
          maxLength: 50
        delivery_address:
          type: string
        notes_commercial:
          type: string
        notes_internal:
          type: string
      required:
      - client_name
      - name
...
    Reinforcement:
      type: object
      properties:
        parent_profile_sku:
          type: string
        reinforcement_sku:
          type: string
          nullable: true
        role:
          type: string
        length_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        qty:
          type: integer
        bay_id:
          type: string
          nullable: true
        leaf_id:
          type: string
          nullable: true
      required:
      - bay_id
      - leaf_id
      - length_mm
      - parent_profile_sku
      - qty
      - reinforcement_sku
      - role
...
    SuccessorRequestRequest:
      type: object
      properties:
        confirmed:
          type: boolean
      required:
      - confirmed
...
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT